[css-in-js] EmptyState widget
